requirements.txt
# Core dependencies
Flask==2.3.2
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1

# Database
SQLAlchemy==2.0.19
python-dotenv==1.0.0

# Utilities
Werkzeug==2.3.6
click==8.1.6
requests==2.31.0

# Async
gevent==23.7.0
gevent-websocket==0.10.1

# Logging & monitoring
python-json-logger==2.0.7

# Data validation
pydantic==2.1.1

# Security
PyJWT==2.8.0
cryptography==41.0.1
